Convert + Clean up new snapshot proposal form (#2322)
* convert to TSX, construct out

* callouts out

* break out helpers and types

* remove div

* rename

* fix import

* styles

* fixed
