+++ packages/commonwealth/client/scripts/views/components/component_kit/cw_radio_group.tsx
@@ -12,7 +12,7 @@ type RadioGroupAttrs = {
   name: string;
   onchange: (e?: any) => void;
   options: Array<RadioButtonType>;
-  toggledOption: string;
+  toggledOption?: string;
 };
 export class CWRadioGroup extends ClassComponent<RadioGroupAttrs> {
   view(vnode: m.Vnode<RadioGroupAttrs>) {
